+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added sessionID to websocket connection
 
 ## [v1.5.1]
 - Added automated releases using travis [#444](https://github.com/xmidt-org/webpa-common/pull/444)